Using BrowserRouter, Route, NavLink, HOC: withRouter, Relative Link, Absolute Link, Styling active Route
