+++ src/components/Post/Post.js
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import './Post.css';
 
 const post = (props) => (
@@ -10,5 +9,4 @@ const post = (props) => (
         </div>
     </article>
 );
-
 export default post;
