+++ src/containers/Blog/Blog.js
@@ -1,63 +1,34 @@
 import React, { Component } from 'react';
 //import axios from 'axios';
-import axios from '../../axios';
-
-import Post from '../../components/Post/Post';
-import FullPost from '../../components/FullPost/FullPost';
-import NewPost from '../../components/NewPost/NewPost';
+import {Route, NavLink} from 'react-router-dom';
 import './Blog.css';
+import Posts from "./Posts/Posts";
+import NewPost from './NewPost/NewPost';
 
 class Blog extends Component {
-	state={
-		posts:[],
-		selectedPostId: null,
-		error: false
-	}
-
-	componentDidMount () {
-		axios.get('/posts')
-			.then( response => {
-				const posts = response.data.slice(0, 4);
-				const updatedPosts = posts.map( post => {
-					return {
-						...post,
-						author: 'Luong'
-					};
-				});
-				this.setState({posts : updatedPosts});
-				//console.log(response);
-			})
-			.catch((error)=>{
-				this.setState({error: true});
-			});
-	}
-
-	postSelectedHandler = (id) => {
-		this.setState({selectedPostId: id});
-	}
-
 	render () {
-		let posts = <p style={{textAlign: 'center'}}> Something went wrong!</p>;
-		if (!this.state.error) {
-			posts = this.state.posts.map(post => {
-				return <Post 
-					clicked={() => this.postSelectedHandler(post.id)} 
-					key={post.id} 
-					title={post.title} 
-					author={post.author}/>
-			});
-		} 
 		return (
-				<div>
-						<section className="Posts">
-								{posts}
-						</section>
-						<section>
-								<FullPost id={this.state.selectedPostId} />
-						</section>
-						<section>
-								<NewPost />
-						</section>
+				<div className="Blog">
+					<header>
+						<nav>
+							<ul>
+								<li><NavLink to="/" 
+											exact
+											activeClassName="my-active"
+											activeStyle={{
+												color: '#fa923f',
+												textDecoration: 'underline'
+											}} >Home</NavLink></li>
+								<li><NavLink to={{
+									pathname: '/new-post',
+									hash: '#submit',
+									search: '?quick-submit=true'
+								}} >New Post</NavLink></li>
+							</ul>
+						</nav>
+					</header>
+					<Route path="/" exact component={Posts} />
+					<Route path="/new-post" exact component={NewPost} />
 				</div>
 		);
 	}
