+++ src/containers/Blog/NewPost/NewPost.js
@@ -10,6 +10,11 @@ class NewPost extends Component {
         author: 'Luong'
     }
 
+    componentDidMount () {
+		console.log(this.props);
+	}
+
+    
     postDataHandler = () => {
       const data = {
 				title: this.state.title,
